fix(release): close Alpha runtime source gates
Goal: 2026-08-09-kungfu-alpha2-buildchain-v3-trusted-release
