+++ docs/qualification/evidence/kungfu-temporal-provenance-cutover.json
@@ -1,8 +1,8 @@
 {
   "acceptedContractDigest": "sha256:13c4679c4ac8764c85e29693bfb59099e21e9786cc6082552198d39393467490",
-  "admissionFactSetRoot": "sha256:57f1160987308819ab114dfbc05edd4493ebb2690ea0d59f3809d1d161c56b33",
-  "admissionPathReceiptRoot": "sha256:9b712617007818556b1277a59ebb2d88f4d2fdcf53eab7932191145067f8cc3b",
-  "admissionReceiptRoot": "sha256:0d63bda8da323b1eea1c59808241d14bfd2beed25c92a32e8ce24445ff0828e4",
+  "admissionFactSetRoot": "sha256:6b28d433e0d56897f5f65b0afc07593f8d8e0c24816759ae924c1c3b156fba20",
+  "admissionPathReceiptRoot": "sha256:bdd9ddd408ce1fca04aad9f9b3b14c3b1aed859e85ce3380c6440dfedac4f4ca",
+  "admissionReceiptRoot": "sha256:8c493a2b971206c252a28096269edc1443f94238a40e35e52348cb1f7fbda24b",
   "ancestrySemanticAuthority": false,
   "artifactCount": 44,
   "artifactSetRoot": "sha256:46169f7d385e5cc296352578cafd29f66762d73b6301f3180ce59560d9f5eef3",
@@ -17,7 +17,7 @@
   "candidateProvenanceObjectRoot": "sha256:eac7fb65b542a951dda314b736fc27fafd86b6d819344510a7ce1802f0a82cdd",
   "candidateRegularFileCount": 14,
   "capsuleRoot": "sha256:8bbc2c4439a99970760cd4f62bb76e55a1323923fa2951aa1985865ab8010dd4",
-  "currentContractDigest": "sha256:8e565f9ac5146d5dceafc3da6b267147fb412937db979bf35a0429966da82197",
+  "currentContractDigest": "sha256:521394028d66ab5985da4c43f4fed46cdb04b838a3b403d3712073dd59858b78",
   "externalPublicationClaimed": false,
   "heavyRebuildPerformed": false,
   "inputUnchanged": true,
@@ -34,7 +34,7 @@
     }
   ],
   "promotionProvenanceObjectRoot": "sha256:dcca03a618d7c493a7fa06c93acc1f38e9d1d4ee799e89ca57a92eb1f7825df4",
-  "qualificationRoot": "sha256:81a6dc46bf72db9c50f92e6df7175275496a00f6b283c95b5cc899d51fb94605",
+  "qualificationRoot": "sha256:2185378adc698a77e57c003d898ceab5bbef8073c9c8a380d6fe794514f32d29",
   "schema": "kungfu.temporal-provenance-cutover-qualification/v1",
   "sourceSha": "ad7c7db6df076f969c5939728bcbe70ccd4771b3",
   "sourceTree": "67a93b5831596555e7c29104421de3a0b97eb865",
